Week 13 - Code To Post
Most of the site is intact. I need to figure out a different solution
to my Blockquote. I also need to get the Horses for Sale page working
right. I am using a template with Javascript for it and need to
decipher the code more so that I can plug it in correctly.